requirements-test.txt
-r requirements.txt

cov-core==1.7
coverage==3.7
flake8==2.1.0
mccabe==0.2.1
pep8==1.4.6
py==1.4.18
pytest==2.4.2
pyflakes==0.7.3
pytest-cov==1.6
pytest-django==2.4
